+++ sphinx/application/data/concepts/repositories/concept-repository-feed/src/main/java/chat/sphinx/concept_repository_feed/FeedRepository.kt
@@ -46,10 +46,13 @@ interface FeedRepository {
 
     fun updatePlayedMark(feedItemId: FeedId, played: Boolean)
 
+    fun updateLastPlayed(feedId: FeedId)
+
     fun getPlayedMark(feedItemId: FeedId): Flow<Boolean?>
 
-    val recommendationsToggleStateFlow: MutableStateFlow<Boolean>
     fun setRecommendationsToggle(enabled: Boolean)
+
+    val recommendationsToggleStateFlow: MutableStateFlow<Boolean>
     val recommendationsPodcast: MutableStateFlow<Podcast?>
 
     suspend fun toggleFeedSubscribeState(feedId: FeedId, currentSubscribeState: Subscribed)