+++ sphinx/screens-detail/podcast-player/podcast-player/src/main/java/chat/sphinx/podcast_player/ui/adapter/PodcastEpisodesListAdapter.kt
@@ -1,5 +1,6 @@
 package chat.sphinx.podcast_player.ui.adapter
 
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
@@ -82,8 +83,10 @@ internal class PodcastEpisodesListAdapter(
                 val same: Boolean = old.playing == new.playing &&
                                     old.downloaded == new.downloaded &&
                                     old.played == new.played &&
-                                    viewModel.isFeedItemDownloadInProgress(old.id) == viewModel.isFeedItemDownloadInProgress(new.id)
-
+                                    old.durationMilliseconds == new.durationMilliseconds &&
+                                    old.currentTimeSeconds == new.currentTimeSeconds &&
+                                    viewModel.isFeedItemDownloadInProgress(old.id) == viewModel.isFeedItemDownloadInProgress(new.id) &&
+                                    viewModel.isEpisodeSoundPlaying(old) == viewModel.isEpisodeSoundPlaying(new)
 
                 if (sameList) {
                     sameList = same
@@ -117,9 +120,11 @@ internal class PodcastEpisodesListAdapter(
                 }
 
                 if (viewState is PodcastPlayerViewState.MediaStateUpdate) {
-                    if (viewState.state is MediaPlayerServiceState.ServiceActive.MediaState.Paused ||
-                        viewState.state is MediaPlayerServiceState.ServiceActive.MediaState.Ended)
-                    {
+                    if (
+                        viewState.state is MediaPlayerServiceState.ServiceActive.MediaState.Paused ||
+                        viewState.state is MediaPlayerServiceState.ServiceActive.MediaState.Ended ||
+                        viewState.state is MediaPlayerServiceState.ServiceActive.MediaState.Playing
+                    ) {
                         episodes = viewState.podcast.getEpisodesListCopy()
                     }
                 }
@@ -137,18 +142,9 @@ internal class PodcastEpisodesListAdapter(
                         }.let { result ->
 
                             if (!diff.sameList) {
-                                val firstVisibleItemPosition = layoutManager.findFirstVisibleItemPosition()
-
                                 podcastEpisodes.clear()
                                 podcastEpisodes.addAll(episodes)
                                 result.dispatchUpdatesTo(this@PodcastEpisodesListAdapter)
-
-                                if (
-                                    firstVisibleItemPosition == 0                               &&
-                                    recyclerView.scrollState == RecyclerView.SCROLL_STATE_IDLE
-                                ) {
-                                    recyclerView.scrollToPosition(0)
-                                }
                             }
                         }
                     }
@@ -337,7 +333,7 @@ internal class PodcastEpisodesListAdapter(
                 onStopSupervisor.scope.launch(viewModel.mainImmediate) {
                     delay(100L)
 
-                    if (podcastEpisode.playing) {
+                    if (viewModel.isEpisodeSoundPlaying(podcastEpisode)) {
                         animationViewPlay.playAnimation()
                     } else {
                         animationViewPlay.pauseAnimation()