+++ sphinx/screens/dashboard/dashboard/src/main/java/chat/sphinx/dashboard/ui/feed/FeedRecentlyPlayedViewModel.kt
@@ -1,0 +1,12 @@
+package chat.sphinx.dashboard.ui.feed
+
+import chat.sphinx.wrapper_feed.Feed
+import chat.sphinx.wrapper_podcast.FeedRecommendation
+import kotlinx.coroutines.flow.StateFlow
+
+interface FeedRecentlyPlayedViewModel {
+    val lastPlayedFeedsHolderViewStateFlow: StateFlow<List<Feed>>
+
+    fun recentlyPlayedSelected(feed: Feed)
+
+}