+++ sphinx/service/concepts/concept-service-media-player/src/main/java/chat/sphinx/concept_service_media/MediaPlayerServiceController.kt
@@ -60,5 +60,5 @@ abstract class MediaPlayerServiceController {
 
     abstract suspend fun submitAction(userAction: UserAction)
 
-    abstract fun getPlayingContent(): Pair<String, String>?
+    abstract fun getPlayingContent(): Triple<String, String, Boolean>?
 }